+++ lib/ReactViews/Map/Navigation/Items/AugmentedVirtualityTool.tsx
@@ -119,11 +119,7 @@ export class AugmentedVirtualityRealignController extends MapNavigationItemContr
 
   @computed
   get visible(): boolean {
-    return this.props.augmentedVirtuality.active && this._visible;
-  }
-
-  set visible(value: boolean) {
-    this._visible = value;
+    return this.props.augmentedVirtuality.active && this.visible;
   }
 
   handleClick(): void {
@@ -219,11 +215,7 @@ export class AugmentedVirtualityHoverController extends MapNavigationItemControl
 
   @computed
   get visible(): boolean {
-    return this.props.augmentedVirtuality.active && this._visible;
-  }
-
-  set visible(v: boolean) {
-    this._visible = v;
+    return this.props.augmentedVirtuality.active && this.visible;
   }
 
   handleClick(): void {