+++ lib/ViewModels/CompositeBar/CompositeBarItemController.ts
@@ -1,4 +1,4 @@
-import { computed, observable } from "mobx";
+import { computed, observable, action } from "mobx";
 import ViewerMode from "../../Models/ViewerMode";
 import React from "react";
 
@@ -107,10 +107,10 @@ export abstract class CompositeBarItemController
 
   /**
    * Whether this item is visible on the screen.
-   * @protected
+   * @private
    */
   @observable
-  protected _visible: boolean = true;
+  private _visible: boolean = true;
 
   /**
    * Gets the {@link this._visible}
@@ -120,11 +120,9 @@ export abstract class CompositeBarItemController
     return this._visible;
   }
 
-  /**
-   * Sets the {@link this._visible}
-   */
-  set visible(value: boolean) {
-    this._visible = value;
+  @action
+  setVisible(v: boolean) {
+    this._visible = v;
   }
 
   /**